requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
appnope==0.1.2; sys_platform == 'darwin'
asgiref==3.5.0; python_version >= '3.7'
backcall==0.2.0
boto3==1.21.27
botocore==1.24.28; python_version >= '3.6'
decorator==5.1.1; python_version >= '3.5'
django-cors-headers==3.11.0
django-rest-auth==0.9.5
django-storages==1.12.3
django==4.0.3
djangorestframework==3.13.1
gunicorn==20.1.0
ipython==7.31.0
jedi==0.18.1; python_version >= '3.6'
jmespath==1.0.0; python_version >= '3.7'
matplotlib-inline==0.1.3; python_version >= '3.5'
parso==0.8.3; python_version >= '3.6'
pexpect==4.8.0; sys_platform != 'win32'
pickleshare==0.7.5
pillow==9.0.1
prompt-toolkit==3.0.28; python_full_version >= '3.6.2'
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pygments==2.11.2; python_version >= '3.5'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2022.1
s3transfer==0.5.2; python_version >= '3.6'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2; python_version >= '3.5'
traitlets==5.1.1; python_version >= '3.7'
urllib3==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
wcwidth==0.2.5
